Changed 'changePlayer'
Now it takes just a String instead of a function (String -> String). I originally used a function because it seemed more flexible. Maybe if you look in a mirror, the descriptions are reversed? Then, you use changePlayer reverse description. Or something.

I think 90% of the time, it's easy to just give a new value. Change "alive" to "dead" or whatever.